refactor(encoding): #111 Phase 1c — extract Simple matcher into simple/
Move SuffixStore, WindowEntry, and the Simple MatchGenerator (level 1
`Fastest` backend) out of the match_generator.rs monolith into a new
`encoding::simple` submodule.

Mechanical move — names, fields, derives, method bodies, and the
`#[inline(always)]` annotations are preserved byte-for-byte. The only
intentional API change is `pub(crate)` visibility on the relocated
items (struct + fields + the handful of methods called from in-file
tests still living in match_generator.rs).

`MIN_MATCH_LEN` and `FAST_HASH_FILL_STEP` are opened to `pub(crate)`
so `simple::` can keep referencing them from their original
declaration site without duplication. `match_table::helpers`,
`row`, and `dfast` now import `MatchGenerator` through
`super::simple::MatchGenerator` instead of the old monolith path.

No behaviour change — full lib test suite (381 tests including the
level22 ratio gate) and clippy clean on default + `--no-default-features`.

Part of #111.
