+++ zstd/src/encoding/mod.rs
@@ -9,12 +9,14 @@ pub(crate) mod match_generator;
 pub(crate) mod util;
 
 // `#111` encoder architecture rewrite — Phase 1 (structural split).
-// `cost_model`, `opt`, and `strategy` now host the relocated cost-model
-// types, the optimal-parser plain-data types, and the `HcParseMode`
-// dispatch tag respectively. `bt`, `hc`, and `match_table` are still
-// doc-only placeholders that name the donor helpers and monolith
-// methods slated to move into them in subsequent #111 phase PRs. The
-// rewrite plan is tracked in
+// `cost_model`, `opt`, `strategy`, `dfast`, `row`, and `simple` now
+// host the relocated cost-model types, the optimal-parser plain-data
+// types, the `HcParseMode` dispatch tag, and the Dfast / Row / Simple
+// matchers respectively. `match_table::helpers` hosts the shared
+// match-finder primitives. `bt` and `hc` are still doc-only
+// placeholders that name the donor helpers and monolith methods
+// slated to move into them in subsequent #111 phase PRs. The rewrite
+// plan is tracked in
 // <https://github.com/structured-world/structured-zstd/issues/111> and
 // the per-phase boundary is `perf/post-pr-110-baseline`.
 pub(crate) mod bt;
@@ -24,6 +26,7 @@ pub(crate) mod hc;
 pub(crate) mod match_table;
 pub(crate) mod opt;
 pub(crate) mod row;
+pub(crate) mod simple;
 pub(crate) mod strategy;
 
 mod frame_compressor;
